Refactor: user_reducer에서 auth_reducer 분리
auth_reducer: login, sign in 관련 로직
user_reducer: userboard, userpost관련로직
