+++ src/_reducers/user_reducer.js
@@ -1,10 +1,4 @@
-import {
-  AUTH_USER,
-  GET_USER_DETAIL,
-  GET_USER__ARR,
-  LOGIN_USER,
-  SIGNUP_USER,
-} from '_types/userTypes';
+import { GET_USER_DETAIL, GET_USER__ARR } from '_types/userTypes';
 
 const initState = {
   myData: null,
@@ -15,14 +9,6 @@ const initState = {
 const userReducer = (state = initState, action) => {
   // type마다 다른 것을 switch로 처리
   switch (action.type) {
-    case LOGIN_USER:
-      return { ...state, myData: action.payload };
-    case SIGNUP_USER:
-      console.log(action.payload);
-      return { ...state };
-    case AUTH_USER:
-      console.log(action.payload);
-      return { ...state };
     case GET_USER_DETAIL:
       return { ...state, targetUser: action.payload };
     case GET_USER__ARR:
