+++ src/_types/authType.js
@@ -1,0 +1,3 @@
+export const LOGIN_USER = 'login_user';
+export const SIGNUP_USER = 'signUp_user';
+export const AUTH_USER = 'auth_user';
